% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentiles.R
\name{exposure_percentile}
\alias{exposure_percentile}
\title{Convert exposures to within-cohort percentiles}
\usage{
exposure_percentile(x)
}
\arguments{
\item{x}{Numeric exposure vector.}
}
\value{
Numeric percentiles in \verb{(0, 1]}.
}
\description{
Convert exposures to within-cohort percentiles
}
