% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtbi-package.R
\docType{package}
\name{qtbi-package}
\alias{qtbi}
\alias{qtbi-package}
\title{Quantum Toxic Burden Index}
\description{
Compute the Quantum Toxic Burden Index (QTBI) from multi-exposure panels
using a fixed quantum-inspired entanglement encoder.
}
\section{Typical workflow}{

\itemize{
\item \code{\link[=estimate_qtbi]{estimate_qtbi()}} adds percentile and QTBI columns to your data frame.
\item Optional \code{reference_doses} enable potency-weighted readout at the index step.
Weights are rescaled automatically so weighted QTBI stays on the same
\verb{[0, n]} scale as the unweighted index.
\item \code{\link[=diagnose_qtbi]{diagnose_qtbi()}} summarizes synergy sensitivity and monotonicity.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/january-msemakweli/qtbi}
  \item Report bugs at \url{https://github.com/january-msemakweli/qtbi/issues}
}

}
\author{
\strong{Maintainer}: January G. Msemakweli \email{jmsemak1@jh.edu}

}
\keyword{internal}
